third_party/blink/web_tests/fast/dom/global-event-handlers-expected.txt

This is a testharness.js-based test.
[FAIL] Set HTMLElement.ondragexit
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate HTMLElement.ondragexit
  assert_true: expected true got undefined
[FAIL] Reflect HTMLElement.ondragexit
  assert_equals: Convert to function expected "function" but got "undefined"
[FAIL] Set SVGElement.ondragexit
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate SVGElement.ondragexit
  assert_true: expected true got undefined
[FAIL] Reflect SVGElement.ondragexit
  assert_equals: Convert to function expected "function" but got "undefined"
[FAIL] Set Document.ondragexit
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate Document.ondragexit
  assert_true: expected true got undefined
[FAIL] Set Window.ondragexit
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate Window.ondragexit
  assert_true: expected true got undefined
[FAIL] Set HTMLElement.onsort
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate HTMLElement.onsort
  assert_true: expected true got undefined
[FAIL] Reflect HTMLElement.onsort
  assert_equals: Convert to function expected "function" but got "undefined"
[FAIL] Set SVGElement.onsort
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate SVGElement.onsort
  assert_true: expected true got undefined
[FAIL] Reflect SVGElement.onsort
  assert_equals: Convert to function expected "function" but got "undefined"
[FAIL] Set Document.onsort
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate Document.onsort
  assert_true: expected true got undefined
[FAIL] Set Window.onsort
  assert_equals: Initially null expected (object) null but got (undefined) undefined
[FAIL] Enumerate Window.onsort
  assert_true: expected true got undefined
Harness: the test ran to completion.
